requirements.txt
PySide6
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pewee
